:recycle: (crdt) moved re-used utils in actual-server to separate package
actual-server does not need to import the full actual-app/api package. It can import only the CRDT
stuff.. so I'm extracting it into a new package to reduce the size of actual-server and make the link
between things more transparent.
