+++ packages/crdt/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@actual-app/crdt",
+  "version": "1.0.0",
+  "license": "MIT",
+  "description": "CRDT layer of Actual",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "lint": "eslint .",
+    "build:node": "tsc --p tsconfig.dist.json",
+    "build": "rm -rf dist && yarn run build:node && cp src/proto/sync_pb.d.ts dist/src/proto/"
+  },
+  "dependencies": {
+    "google-protobuf": "^3.12.0-rc.1",
+    "murmurhash": "^0.0.2",
+    "uuid": "3.3.2"
+  },
+  "devDependencies": {
+    "typescript": "^5.0.2"
+  }
+}
