+++ packages/crdt/src/main.ts
@@ -1,0 +1,13 @@
+import * as SyncPb from './proto/sync_pb';
+export {
+  merkle,
+  getClock,
+  setClock,
+  makeClock,
+  makeClientId,
+  serializeClock,
+  deserializeClock,
+  Timestamp,
+} from './crdt';
+
+export const SyncProtoBuf = SyncPb;
