+++ packages/crdt/src/proto/sync_pb.d.ts
