+++ packages/loot-core/src/server/api.ts
@@ -1,3 +1,5 @@
+import { getClock } from '@actual-app/crdt';
+
 import * as connection from '../platform/server/connection';
 import {
   getDownloadError,
@@ -22,7 +24,6 @@ import {
 } from './api-models';
 import { runQuery as aqlQuery } from './aql';
 import * as cloudStorage from './cloud-storage';
-import { getClock } from './crdt';
 import * as db from './db';
 import { runMutator } from './mutators';
 import * as prefs from './prefs';
