+++ packages/loot-core/src/server/aql/schema/executors.test.ts
@@ -1,9 +1,9 @@
+import { setClock } from '@actual-app/crdt';
 import fc from 'fast-check';
 
 import * as arbs from '../../../mocks/arbitrary-schema';
 import query from '../../../shared/query';
 import { groupById } from '../../../shared/util';
-import { setClock } from '../../crdt';
 import * as db from '../../db';
 import { batchMessages, setSyncingMode } from '../../sync/index';
