+++ packages/loot-core/src/server/sync/index.ts
@@ -1,3 +1,11 @@
+import {
+  serializeClock,
+  deserializeClock,
+  getClock,
+  Timestamp,
+  merkle,
+} from '@actual-app/crdt';
+
 import { captureException } from '../../platform/exceptions';
 import * as asyncStorage from '../../platform/server/asyncStorage';
 import * as connection from '../../platform/server/connection';
@@ -5,13 +13,6 @@ import logger from '../../platform/server/log';
 import { sequential, once } from '../../shared/async';
 import { setIn, getIn } from '../../shared/util';
 import { triggerBudgetChanges, setType as setBudgetType } from '../budget/base';
-import {
-  serializeClock,
-  deserializeClock,
-  getClock,
-  Timestamp,
-  merkle,
-} from '../crdt';
 import * as db from '../db';
 import { PostError, SyncError } from '../errors';
 import app from '../main-app';
